cmake: modify the link attribute to private to avoid re-initialization
The dpdk library initializes the EAL using constructors and global
variables, and cannot be re-initialized. Using the default
target_link_libraries property, both application and libceph-common.so have
a DPDK constructor that causes repeated initialization.Modify the link
attribute to private so that there is only one constructor in libcep-common.

Signed-off-by: Chunsong Feng <fengchunsong@huawei.com>