msg/dpdk: add shared library to fix dupplicate DPDK EAL link issue
The common_async_dpdk library depends on the dpdk rte_eal.a file.
If the static library is used, the dependency is transferred to
libceph-common. By default, libceph-common uses the PUBLIC keyword,
the keyword PUBLIC determines that both the application and
libceph-common link to dpdk rte_eal.a. As a result, the global variable
in dpdk rte_eal.a has two copies. This patch provides both static and
shared libraries.let libcommon.a link to static library and libceph-common.so
link to dynamic library.

Fixes: https://tracker.ceph.com/issues/42861

Signed-off-by: Chunsong Feng <fengchunsong@huawei.com>
Signed-off-by: luo rixin <luorixin@huawei.com>